Added new (pseudo-like code, thus commented out, for) JOINing with raw data sets.
The fundamental assumption that Spark sends the actual data, and not just pointers, to each worker node seems to be wrong (this would require the driver program to load all data which 1) there would be no reason for the driver program to do and 2) would be inefficient because it would not be done in parallell).

Also, added Scala "Hello World!".